+++ tests/testthat/test-as_tibble.R
@@ -15,3 +15,13 @@ test_that("as_tibble() works", {
     tibble(x1 = c("a", "a", "b"), x2 = c("1", "2", "3"))
   )
 })
+
+test_that("as_tibble() on grouped input returns grouped tibble", {
+  test_df <- tibble(x1 = c("a", "a", "b"), x2 = 1:3)
+  test_pl <- as_polars_df(test_df)
+
+  expect_equal(
+    test_pl |> group_by(x1) |> mutate(x3 = x2 + 1) |> as_tibble(),
+    test_df |> group_by(x1) |> mutate(x3 = x2 + 1)
+  )
+})